+++ spec/dummy/app/graphs/edges/has_job.rb
@@ -1,8 +1,11 @@
 module Edges
-  class WorksAt
+  class HasJob
     include ApacheAge::Edge
 
     attribute :employee_role, :string
+    attribute :start_node, :person
+    attribute :end_node, :company
+
     validates :employee_role, presence: true
   end
 end